requirements.txt
Flask
qrcode
Pillow
reportlab
Werkzeug
gunicorn